+++ packages/plugin-audio/web/src/components/AudioImage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from '../../statics/styles/audio.scss';
+import styles from '../../statics/styles/audio.rtlignore.scss';
 import classNames from 'classnames';
 import { Image } from 'wix-rich-content-ui-components';
 import { getImageSrc } from 'wix-rich-content-common/libs/imageUtils';