+++ packages/plugin-audio/web/src/components/ContextMenu.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from '../../statics/styles/audio.scss';
+import styles from '../../statics/styles/audio.rtlignore.scss';
 import { ClickOutside } from 'wix-rich-content-editor-common';
 import DropdownMenu from './DropdownMenu';
 import { ContextMenuIcon } from 'wix-rich-content-ui-components';