+++ packages/plugin-audio/web/src/createAudioPlugin.ts
@@ -2,7 +2,7 @@ import createToolbar from './toolbar/createToolbar';
 import { Component, DEFAULTS } from './audio-component';
 import type { AudioPluginEditorConfig } from './types';
 import { AUDIO_TYPE } from './types';
-import { createBasePlugin } from 'wix-rich-content-plugin-commons';
+import { createBaseMediaPlugin, createBasePlugin } from 'wix-rich-content-plugin-commons';
 import type { CreatePluginFunction } from 'wix-rich-content-common';
 
 const createAudioPlugin: CreatePluginFunction<AudioPluginEditorConfig> = config => {
@@ -10,7 +10,7 @@ const createAudioPlugin: CreatePluginFunction<AudioPluginEditorConfig> = config 
   const disableDownload = config?.uiSettings?.disableDownload;
   const type = AUDIO_TYPE;
   return createBasePlugin({
-    component: Component,
+    component: createBaseMediaPlugin(Component),
     type,
     toolbar: createToolbar({
       t,